separate patientQueryCreator.js
