+++ src/operations/common/securityTagManager.js
@@ -4,8 +4,8 @@ const { ScopesManager } = require('../security/scopesManager');
 const { AccessIndexManager } = require('./accessIndexManager');
 const { SecurityTagSystem } = require('../../utils/securityTagSystem');
 const { PatientFilterManager } = require('../../fhir/patientFilterManager');
-const { isUuid } = require('../../utils/uid.util');
 const { FieldMapper } = require('../query/filters/fieldMapper');
+const { R4SearchQueryCreator } = require('../query/r4');
 
 /**
  * This class manages queries for security tags
@@ -16,8 +16,9 @@ class SecurityTagManager {
      * @param {ScopesManager} scopesManager
      * @param {AccessIndexManager} accessIndexManager
      * @param {PatientFilterManager} patientFilterManager
+     * @param {R4SearchQueryCreator} r4SearchQueryCreator
      */
-    constructor ({ scopesManager, accessIndexManager, patientFilterManager }) {
+    constructor ({ scopesManager, accessIndexManager, patientFilterManager, r4SearchQueryCreator }) {
         /**
          * @type {ScopesManager}
          */
@@ -35,6 +36,12 @@ class SecurityTagManager {
          */
         this.patientFilterManager = patientFilterManager;
         assertTypeEquals(patientFilterManager, PatientFilterManager);
+
+        /**
+         * @type {R4SearchQueryCreator}
+         */
+        this.r4SearchQueryCreator = r4SearchQueryCreator;
+        assertTypeEquals(r4SearchQueryCreator, R4SearchQueryCreator);
     }
 
     /**
@@ -66,34 +73,6 @@ class SecurityTagManager {
     }
 
     /**
-     * if there is already an $and statement then just add to it
-     * @param {import('mongodb').Document} query
-     * @param {import('mongodb').Document} andQuery
-     * @return {import('mongodb').Document}
-     */
-    appendAndQuery (query, andQuery) {
-        if (query.$and) {
-            query.$and.push(
-                andQuery
-            );
-            return query;
-        } else if (Object.keys(query).length === 0) { // empty query then just replace
-            return {
-                $and: [
-                    andQuery
-                ]
-            };
-        } else {
-            return {
-                $and: [
-                    query,
-                    andQuery
-                ]
-            };
-        }
-    }
-
-    /**
      * returns the passed query by adding a check for security tgs
      * @param {string} resourceType
      * @param {string[]} securityTags
@@ -103,8 +82,8 @@ class SecurityTagManager {
      * @return {import('mongodb').Document}
      */
     getQueryWithSecurityTags ({
-        resourceType, securityTags, query, useAccessIndex = false, useHistoryTable
-    }) {
+                                 resourceType, securityTags, query, useAccessIndex = false, useHistoryTable
+                             }) {
         const fieldMapper = new FieldMapper({ useHistoryTable });
         if (securityTags && securityTags.length > 0) {
             let securityTagQuery;
@@ -150,118 +129,7 @@ class SecurityTagManager {
             }
 
             // if there is already an $and statement then just add to it
-            query = this.appendAndQuery(query, securityTagQuery);
-        }
-        return query;
-    }
-
-    /**
-     * Gets Patient Filter Query
-     * @param {string[] | null} patientIds
-     * @param {import('mongodb').Document} query
-     * @param {string} resourceType
-     * @param {boolean} useHistoryTable
-     * @return {import('mongodb').Document}
-     */
-    getQueryWithPatientFilter ({ patientIds, query, resourceType, useHistoryTable }) {
-        if (!this.patientFilterManager.canAccessResourceWithPatientScope({ resourceType })) {
-            throw new ForbiddenError(`Resource type ${resourceType} cannot be accessed via a patient scope`);
-        }
-        const fieldMapper = new FieldMapper({ useHistoryTable });
-        // separate uuids from non-uuids
-        const patientUuids = patientIds.filter(id => isUuid(id));
-        let patientsUuidQuery, patientsNonUuidQuery;
-        if (patientUuids && patientUuids.length > 0) {
-            const inQuery = {
-                $in: resourceType === 'Patient' ? patientUuids : patientUuids.map(p => `Patient/${p}`)
-            };
-            /**
-             * @type {string|string[]|null}
-             */
-            const patientFilterProperty = this.patientFilterManager.getPatientPropertyForResource({
-                resourceType
-            });
-            if (patientFilterProperty) {
-                if (Array.isArray(patientFilterProperty)) {
-                    patientsUuidQuery = {
-                        $or: patientFilterProperty.map(p => {
-                                // if patient itself then search by _uuid
-                                if (p === 'id') {
-                                    return { [fieldMapper.getFieldName('_uuid')]: inQuery };
-                                }
-                                return {
-                                    [fieldMapper.getFieldName(p.replace('.reference', '._uuid'))]: inQuery
-                                };
-                            }
-                        )
-                    };
-                } else {
-                    // if patient itself then search by _uuid
-                    if (patientFilterProperty === 'id') {
-                        patientsUuidQuery = { [fieldMapper.getFieldName('_uuid')]: inQuery };
-                    } else {
-                        patientsUuidQuery = {
-                            [
-                                fieldMapper.getFieldName(
-                                    patientFilterProperty.replace('.reference', '._uuid')
-                                )
-                            ]: inQuery
-                        };
-                    }
-                }
-            }
-        }
-        const patientNonUuids = patientIds.filter(id => !isUuid(id));
-        if (patientNonUuids && patientNonUuids.length > 0) {
-            const inQuery = {
-                $in: resourceType === 'Patient' ? patientNonUuids : patientNonUuids.map(p => `Patient/${p}`)
-            };
-            /**
-             * @type {string|string[]|null}
-             */
-            const patientFilterProperty = this.patientFilterManager.getPatientPropertyForResource({
-                resourceType
-            });
-            if (patientFilterProperty) {
-                if (Array.isArray(patientFilterProperty)) {
-                    patientsNonUuidQuery = {
-                        $or: patientFilterProperty.map(p => {
-                                // if patient itself then search by _sourceId
-                                if (p === 'id') {
-                                    return { [fieldMapper.getFieldName('_sourceId')]: inQuery };
-                                }
-                                return {
-                                    [fieldMapper.getFieldName(p.replace('.reference', '._sourceId'))]: inQuery
-                                };
-                            }
-                        )
-                    };
-                } else {
-                    // if patient itself then search by _sourceId
-                    if (patientFilterProperty === 'id') {
-                        patientsNonUuidQuery = { [fieldMapper.getFieldName('_sourceId')]: inQuery };
-                    } else {
-                        patientsNonUuidQuery = {
-                            [
-                                fieldMapper.getFieldName(
-                                    patientFilterProperty.replace('.reference', '._sourceId')
-                                )
-                            ]: inQuery
-                        };
-                    }
-                }
-            }
-        }
-        let patientsQuery;
-        if (patientsUuidQuery && patientsNonUuidQuery) {
-            patientsQuery = {
-                $or: [patientsUuidQuery, patientsNonUuidQuery]
-            };
-        } else if (patientsUuidQuery || patientsNonUuidQuery) {
-            patientsQuery = patientsUuidQuery || patientsNonUuidQuery;
-        }
-        if (patientsQuery) {
-            query = this.appendAndQuery(query, patientsQuery);
+            query = this.r4SearchQueryCreator.appendAndQuery(query, securityTagQuery);
         }
         return query;
     }
